django_rest_main/requirements.txt
Django==4.2.27
djangorestframework==3.14.0
django-filter==23.3
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
django-mailjet==0.3.1

